+++ packages/orq-rc/src/models/operations/createconversation.ts
@@ -270,7 +270,7 @@ export const CreateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kfdy64ra24s3wr7fgbxzh36y"),
+  _id: z.string().default("conv_01kfe1mwtc5jzmh1rveatwt3ms"),
   kind: CreateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -302,7 +302,7 @@ export const CreateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kfdy64ra24s3wr7fgbxzh36y"),
+  id: z.string().default("conv_01kfe1mwtc5jzmh1rveatwt3ms"),
   kind: CreateConversationKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),