+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -576,7 +576,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KFDY65K88EQYPSFX567FECWQ"),
+  _id: z.string().default("01KFE1MXAKT8YB63RVW4BT69N7"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,
@@ -619,7 +619,7 @@ export const CreateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KFDY65K88EQYPSFX567FECWQ"),
+  id: z.string().default("01KFE1MXAKT8YB63RVW4BT69N7"),
   displayName: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$outboundSchema,