+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -316,7 +316,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kfdy64rj74ptt2669tfbkqvx"),
+  _id: z.string().default("conv_01kfe1mwtk9pf9zrs7r8esh3b5"),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -348,7 +348,7 @@ export const UpdateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kfdy64rj74ptt2669tfbkqvx"),
+  id: z.string().default("conv_01kfe1mwtk9pf9zrs7r8esh3b5"),
   kind: UpdateConversationKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),